Moved tests to main folder
