+++ __tests__/integration/websocket/main/serverless.yml
@@ -14,7 +14,7 @@
 service: integration-tests-websocket-main
 
 plugins:
-  - ./../../..
+  - ./../../../..
 
 provider:
   name: aws
